requirements.txt
flask
flask-restx
python-dotenv
sqlalchemy
flask-cors
pymongo
bcrypt
pandas
pyjwt
psycopg2